[step3] refactor: 상세 도메인 패키지 분리
