+++ src/main/kotlin/domain/MinePositionGenerator.kt
@@ -1,5 +1,7 @@
 package domain
 
+import domain.position.Positions
+
 interface MinePositionGenerator {
     fun generate(): Positions
 }
