+++ src/main/kotlin/domain/cell/Cell.kt
@@ -1,3 +1,5 @@
-package domain
+package domain.cell
+
+import domain.position.Position
 
 data class Cell(val position: Position, val property: CellProperty)
