+++ src/main/kotlin/domain/cell/CellState.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.cell
 
 enum class CellState {
     OPEN, HIDE
