+++ src/main/kotlin/domain/cell/CellType.kt
@@ -1,4 +1,4 @@
-package domain
+package domain.cell
 
 enum class CellType {
     MINE, NORMAL,
