+++ src/test/kotlin/minesweeper/domain/TestMinePositionGenerator.kt
@@ -2,9 +2,9 @@ package minesweeper.domain
 
 import domain.MineCountNumber
 import domain.MinePositionGenerator
-import domain.Position
-import domain.Positions
-import domain.toPositions
+import domain.position.Position
+import domain.position.Positions
+import domain.position.toPositions
 import domain.toPositiveNumber
 
 class TestMinePositionGenerator(private vararg val positionValues: Pair<Int, Int>) : MinePositionGenerator {
@@ -12,7 +12,7 @@ class TestMinePositionGenerator(private vararg val positionValues: Pair<Int, Int
         return positionValues.map { (row, column) -> Position.of(row, column) }.toPositions()
     }
 
-    fun getMineCount(height: Int, width:Int): MineCountNumber {
+    fun getMineCount(height: Int, width: Int): MineCountNumber {
         return MineCountNumber(positionValues.size, height.toPositiveNumber(), width.toPositiveNumber())
     }
 }
